+++ src/vscode/treeviews/workspaces/PowerBIParameters.ts
@@ -51,7 +51,7 @@ export class PowerBIParameters extends PowerBIWorkspaceTreeItem {
 			try {
 				let items: iPowerBIDatasetParameter[] = await PowerBIApiService.getItemList<iPowerBIDatasetParameter>(this.apiPath);
 				for (let item of items) {
-					let treeItem = new PowerBIParameter(item, this.groupId, this);
+					let treeItem = new PowerBIParameter(item, this.parent.id, this);
 					children.push(treeItem);
 					PowerBICommandBuilder.pushQuickPickItem(treeItem);
 				}