+++ src/vscode/treeviews/workspaces/PowerBIWorkspacesTreeProvider.ts
@@ -55,7 +55,9 @@ export class PowerBIWorkspacesTreeProvider implements vscode.TreeDataProvider<Po
 		if (showInfoMessage) {
 			vscode.window.showInformationMessage('Refreshing Workspaces ...');
 		}
-		await ThisExtension.initialize(ThisExtension.extensionContext);
+		if(!item) { // on the root we re-initialize the connection
+			await ThisExtension.initialize(ThisExtension.extensionContext);
+		}
 		this._onDidChangeTreeData.fire(item);
 	}